Update Avatar extension (#1923)
* Update Avatar extension

* Fix bug

* Fix bug

* Fig Advanced Options bug

* Update tips

* Update tips

* Update package.json

* Update CHANGELOG.md
