+++ extensions/avatar/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Avatar Changelog
 
+## [Add new avatar layout options] - 2022-06-09
+
+- Add new avatar layout options: List or Grid.
+
 ## [Initial Version] - 2022-06-06
 
 - An avatar library for designers and developers.
