+++ extensions/avatar/package.json
@@ -21,9 +21,51 @@
       "mode": "view"
     }
   ],
+  "preferences": [
+    {
+      "name": "layout",
+      "type": "dropdown",
+      "description": "Set the layout of the avatar items.",
+      "required": true,
+      "title": "Avatar Layout",
+      "default": "Grid",
+      "data": [
+        {
+          "title": "Grid",
+          "value": "Grid"
+        },
+        {
+          "title": "List",
+          "value": "List"
+        }
+      ]
+    },
+    {
+      "name": "itemSize",
+      "type": "dropdown",
+      "description": "Set the size of the avatar items. (Only works with Grid layout)",
+      "required": false,
+      "title": "Avatar Item Size",
+      "default": "small",
+      "data": [
+        {
+          "title": "Small",
+          "value": "small"
+        },
+        {
+          "title": "Medium",
+          "value": "medium"
+        },
+        {
+          "title": "Large",
+          "value": "large"
+        }
+      ]
+    }
+  ],
   "dependencies": {
     "@multiavatar/multiavatar": "^1.0.7",
-    "@raycast/api": "^1.35.1",
+    "@raycast/api": "^1.36.0",
     "axios": "^0.27.2",
     "d3-color": "^3.1.0",
     "fs-extra": "^10.1.0",
