+++ extensions/avatar/src/components/action-open-command-preferences.tsx
@@ -1,0 +1,14 @@
+import { Action, ActionPanel, Icon, openCommandPreferences } from "@raycast/api";
+
+export function ActionOpenCommandPreferences() {
+  return (
+    <ActionPanel.Section>
+      <Action
+        icon={Icon.Gear}
+        title="Open Command Preferences"
+        shortcut={{ modifiers: ["cmd"], key: "," }}
+        onAction={openCommandPreferences}
+      />
+    </ActionPanel.Section>
+  );
+}
