+++ extensions/avatar/src/types/preferences.ts
@@ -1,0 +1,4 @@
+export interface Preferences {
+  layout: string;
+  itemSize: string;
+}
